Fix all RemovedInMarshmallow4Warning warnings
- replace marshmallow.fields with cmk.utils.fields when metadata
warnings is thrown
- update explicit args to new keywords

Check that all warnings are removed by adding the pytest flag
`-W error::DeprecationWarning:marshmallow.fields`.

Unfortunately we cannot configure filterwarnings with the same
effect in the pytest config, see [1] for some hints why this
might be. My assumption is that the -W flag is passed directly
to the python interpreter before pytest does it's own warning setup.

[1] https://nedbatchelder.com/blog/201810/why_warnings_is_mysterious.html

Change-Id: Ib66d091e738ff3ee8a938ee72aea0073df62c1e1