Autocompleter fixes
All sites auto completer is the empty choice. It is more informative than
the empty string for not being set. It is even more helpful on Grafana
where The filter is always active.

Fix Graph template default title for single metric automatic graphs. The
"str" wrap to combat lazy strings plays against the dictionary non value
giving the string "None", that is not a title

Change-Id: I98875ac83f378b74622e872a7958de70814ef301