Fix waiting for configuration activation during integration tests
Previously the function activate_changes_and_wait_for_completion
was not really waiting. The waiting code was in place, but the API
was triggered in a way where it did not trigger the redirect to the
wait-for-activation endpoint. Instead, the activation was started
and returning back.

All this did only rarely result in failures, because the
activate_changes_and_wait_for_core_reload function was waiting for the
core to reload during the activation. In many cases this was "good
enough", because the activation procedure was finishing fast enough
after the core restart was recognized.

In load situations this resulted in unexpected inconsistencies which
then made test jobs fail as a consequence.

Also add some more debug output to be able to track down similar issues
more easily.

Change-Id: If81042a00ad56c3aa662ea61f615ee0888ade3b6