13615 FIX Add margin to Graph titles in reports
Some graphs like "bandwith" display the unit of the vertical axis on the top
left coner of the graph instead of next to the values on the axis.

Werk 12998 resulted in an overlap between the unit of the graph and its title.
With this Werk an extra margin is added to the title on those graphs.

Graphs that use "graph_unit", display the order of magnitude with its
unit for all values on the graph. That information is on the top left
corner and conflicts with the title. Because those graphs are already
and exception, add the extra margin to the title only to those graphs.

SUP-8896

Change-Id: I099b162807a095d28e2a7770f2e5cdf4c8ccbcff