Encrypter: Move base64 encoding back to ValueSpec
The encoding is only required for the HTML rendering:  This is
the responsibility of the ValueSpec as can be seen in
ImageUpload and SSHKeyPair.

This lets us further simplify the Encrypter and
PasswordStoreObfuscater and removes the requirement for the C
plugins to decode the base64 encoding.

The new typing lets us use an actual number for the protocol
version instead of an untyped "int-string".  We can also remove
the type checks from the unit tests (we use mypy here) and the
equality checks between encrypted and clear-text: there cannot
be equality anymore because encrypted and clear-text have
different types.

CMK-9563

Change-Id: I3219186ac0f72650e70ac28fe5b6f0524befe280