Add resource discovery to gcp special agent
GCP metrics are stored in a more modern timeseries database. I will only
receive data if the database has entries in the requested interval. I.e.
if I do not download objects from a bucket the send_bytes_count metric
will not contain any values, we default to 0 in the check plugin.

For buckets this is still OK because the object_count metric is always
returned. Note this is only true 24h after creation of the bucket, as
this metric is only updated once every 24h.

Other gcp resources like cloud functions however do not return
any metric if the function is not used. Now I can just collect
information on the cloud functions and add them to the section.

Change-Id: I5337df11ee0d2e32f04799299552375e229b97f6