13746 Support Diagnostics: Extend and improve functionality
The usage of the GUI was improved. There's a separate section "Checkmk Log files" now, the component "Business Intelligence" was added and the inline help now contains a description of each file.
The sctructure of the tar.gz file was slightly changed and the following information is now always included in the dump: Checkmk performance data (number of helper processes, current core statistics), HW information (CPU and Memory), and the environment variables of the site user.

Change-Id: I9d2bcbbd84c4aef0df5ee69d6a801272f045eb1d