Python 3 modules: Use standard compiler instead of our own
In our CI containers we always use our own compiler chain. This is
mostly useful for building the cmc on all platforms with the same
compiler of our choice. Since we link the cmc statically there is no
compatibility issue with the distro libraries.

But there are 3rd party C++ components in Checkmk, like protobuf and
grpc, which are dynamically linked with the libstc++. To be compatible
with the target distros, we need to use the distros standard compiler to
build them. Only this way we can produce libraries which are compatible
with the local libstdc++.

All this is specific to our build containers and only needs to have an effect there.
See buildscripts/infrastructure/build-nodes/scripts/install-gnu-toolchain.sh

This change fixes the build of the grpc module.

Change-Id: I13dd9631b1cc609ef3fc4b18c196450b98d6cfc8