Clarify arguments to `Encrypter._secret_key`
The function is a wrapper over `hashlib.scrypt`.  Where it is OK
to fix some of its arguments, pulling one parameter out of the
global scope is really confusing.

That should also let us simplify our unit tests and reduce
the need for patching and IO but this is out of scope for
this commit.

CMK-9563

Change-Id: Ic8172306548b638cb9e9532b66fd2e4264322572