13721 SEC Stop returning password hashes over WebAPI
Previous to this Werk the <i>WebAPI</i> returned the password hashes and
SessionIds when calling the <i>get_all_users</i> action. If the user was a
automation user the password was returned in clear-text.

This data should not be returned ever. If you forgot a password you have to
reset it.

Change-Id: I86f1666ac44c3a194bd334089245a5c3e7c6a06f