Improve protobuf performance
When using python protobuf with dynamically linked libraries, it is
working with the same performance as the modules we get from pypi.

To make these libraries work, we need to pack the libstdc++ and libgcc
libraries from the compiler we use in our build containers. The compiler
we use links against libraries that are are newer than the target
distros libraries. To make our executables work as intended, we ship the
required libraries with Checkmk.

Another approach to make it work is linking the binaries and libraries
statically. That was working with the protobuf libraries and protoc, but
was not performant enough and also resulted in larger binaries. Have a
look at the ticket below for further information.

Centralizing the GCC version definition along the way.

CMK-9522

Change-Id: I9fc416e164757d0f7daa0d49a6edc24a4f15f23a