12333 FIX Disabled  possibility to add too big files to the agent installer
Previously, the limit of 20MB for any file packed by the bakery in the
installer could be enforced only during installation of the agent. This
could cause a serious problem during mass updating of multiple monitored
hosts: the agent installation process will fail on every host to be updated.

Since this release Bakery forbids adding too big files to the agent installer
thus eliminating above mentioned problem with installation and update of the
agent.

Change-Id: Ibf07c455481c32944912bb9c2cf09e0e794511aa