Fix breadcrumb rendering
If a element in the breadcrumbs contain HTML code it broke the rendering
of the breadcrumbs line.

The title was passed to render_element which escapes only permissve,
allowing tags such as `<h1>` or `<br>`. Now the titles are escaped
stricter which keeps the breadcrumbs line prettier but displays the
HTML-tags.

Change-Id: I97f468cc21aa9a9da9605dfb6cc3d6981e5852ec