Update config received from old major version central site
This ensures mixed setups of the previous major central site
and newer remote sites work. The remote site needs to have a
configuration which is compatible with it's version.

Previously 'cmk-update-config' was only executed when processing a
snapshot that was created with a 1.6 central site. But now, with Checkmk
2.1 in a remote site, we are not dealing with 1.6 central sites anymore
but with 2.0 central sites.

Instead of dealing with specific version numbers here, we only check
whether or not the major version of the central and remote site is
exactly the same. If not, then 'cmk-update-config' is executed.

A full compatibility check of the sites is not necessary here. It is
done in the general automation call handling.

Also added some debug logging along the way.

Change-Id: I3c8cc581a3cbd9c308c210681fe0370693169c58