Do not create changes during cmk-update-config
During the executing of the GUI configuration transformation, the
changes recording mechanism of the Setup was creating change entries.
These did not really cause a problem for standalone or central sites in
distributed environments.

But for remote sites, this could cause problems, because the resulting
pending changes prevented future synchronizations of configuration
changes.

CMK-9467

Change-Id: Id2288bfbe03c9113060079a6249f2194d914da89