Make errors visible in parser + fix
`cmk.utils.debug.debug_mode` must be set explicitly for the unit tests
or the errors might no be visible.  This is what happened in the agent
parser.

This patch properly sets debug mode and fixes code and tests until they
are green again.  We added some more tests on the way.

CMK-9820

Change-Id: I6886b87604da1f5d62982a70d5e696373f6a5086