Do not call instance methods from __init__
It is not correct to call instance methods from `__init__()` because
they take `self` as their first argument but `self` is not initialized
until after `__init__` is finished.

This can result in strange error messages such as:

 self = <[AttributeError("'SectionName' object has no attribute '_value'")
    raised in repr()] SectionName object at 0x7f760dd276d0>

What we *can* do instead is calling class methods.  Now, instead of
calling class methods with `type(self)` in `__init__`, the more proper
way is to validate the arguments in `__new__` and keep `__init__` as
simple as possible.

Furthermore, we can see from the code that this is what was expected as
`_parse_valid()` (now `_validate_args()`) was always acting on
`self.__class__` and never on `self` and `self.__class__`, or
`type(class)` *is* `cls` in a class method.

Change-Id: I8ad0d717fb96719de075d808e0383b19d4782525