sles-15sp3: Fix rpm compatibility
The image has rpm-ndb installed with a newer version than our mirror
holds. To make the installation procedure downgrade to the expected
version, we also install these packages with the force-resolution flag.

The error was:

------------
Problem: the to be installed rpm-build-4.14.1-29.46.x86_64 requires 'rpm = 4.14.1', but this requirement cannot be provided
  not installable providers: rpm-ndb-4.14.1-29.2.x86_64[sles15sp3]
                   rpm-4.14.1-29.46.x86_64[sles15sp3]
 Solution 1: downgrade of rpm-ndb-4.14.3-37.2.x86_64 to rpm-ndb-4.14.1-29.2.x86_64
 Solution 2: do not install rpm-build-4.14.1-29.46.x86_64
 Solution 3: break rpm-build-4.14.1-29.46.x86_64 by ignoring some of its dependencies
------------

Change-Id: I7a0ebb207dea9ef0d2c8933b4d0aad0bba59e8f5