13767 FIX API: fix show all users endpoint call when users have no emails
In the previous version, the show all users endpoint returned an error
if one of the users had no email present in its contact information.
This werk fixes this behaviour. It should be noted that this case is only
relevant for internal users. Users created through the REST-API or the UI
will always have an email placeholder field even if an email was not
specified.

Change-Id: I7fd0713cf07b95ab084980bede1dc1e94b0f158e