Clarify relationship between Encrypter and _PasswordStoreObfuscator
A class with only classmethods is not a class in the traditional OO
sense but a collection of free functions in a namespace.  Inheriting a
namespace with free functions (aka inheriting functions) is uncommon.
Even if we had a class in the traditional OO sense, implementation
inheritance is usually frowned upon.  So let us remove the inheritance.

The way the API functions: encrypt and decrypt used to cut the header in
the obfuscated password store was also rather counter-intuitive.

Change-Id: I1c703b516feb58d2b0f0ab72cb3c32a6bd7f97c9