13768 API: addition of enforce password change as configurable option for user create endpoint and moving the option in for the update endpoint
In the previous version, the enforce_password_change option was always set to True
when creating a new user with a password and was not a configurable option. This werk
changes this and introduces this as a configurable option defaulting to False. In
addition, this werk also moves this option to the auth password details for the update endpoint.

Change-Id: I6f8742d96f57880304069c5b62b7f4003ba16ab9