13344 KUBE: kube_pod_status updated rule set
The status of a pod with an erronous container will often change. This happens even if there is
no meaningful change in the underlying problem. Previously, the kube_pod_status check would alert
if a specific status would persist for too long. However, changes of status would reset the
corresponding timer. The new rule set allows configuring a timer for a group of statuses.

CMK-9837

Change-Id: Ic2a9a1b9e3861fa58b31b4fe4bc0c0d94ebd1e12