Add service filtering to gcp api integration
To conserve costs we want to limit the number of API calls made to GCP.
For each service the number of API calls is not determined by the number
of gcp resources, i.e. buckets. Instead we only limit the number of
services monitored by the api integration. Any further filtering can be
done in checkmk itself.

Change-Id: If095771e0523f88c6bd486fe93fc89fb542af027