13638 FIX REST API: Two-factor protected accounts may not access API
The REST API authentication was not handling the new two-factor authentication.
Since the authentication with the security key and one time tokens is not
usable with the API, we now completely block access to the REST API for those
user accounts.

If you have your personal account protected with the two-factor authentication
and want to use the REST API, you may create a dedicated (automation) account
for that.

CMK-9898

Change-Id: I4d9bd48eb6a7afb011e8bb968080806ddf217715