13411 FIX Cisco Prime: Add host options
Before this change the ip address was used to connect to the Cisco Prime
server. This could lead to problems if a TLS connection was used and the
certificate was only valid for the domain and not for the ip address.

Now it's possible to choose whether the ip address, the host name or a
custom host should be used.

SUP-9637

Change-Id: Ic6e734c3f04282e617b64b7b81398eed023ecc57