Changes to agent receiver trigger python3-modules rebuild
The agent receiver is packed by the OMD package python3-modules. It is
added to the package through our Pipfile. In the past one had to
increase the build ID manually to invalidate the OMD package cache.

With this change a change to the relevant files is automatically
detected resulting in automatic packaging as one would expect.

The downsite of this packaging approach is, that every time we change
our internal code, the python3-modules package has to be rebuilt which
takes ~30 minutes per distro. But since the result is cached and the
controller is not changed that often, it is not such a big problem.

CMK-9880

Change-Id: I81aeea3695c17d1ff7f865d0cf3db389d310ed4d