13412 FIX Update manpage for check 'seh_ports'
This check shows the state of the ports of the SEH myUTN dongle server
devices. The state of the port is not interpreted by checkmk, only displayed.
It is always shown as OK, except when the state changes:

If the current port state differs from the state seen at service
discovery time, the service will report WARN. To acknowledge the changed
port, execute the service discovery again.

SUP-9685

Change-Id: I66fe6f4bf6d77a9ad840b9b72dfb190036851369