13618 FIX Checkmk free edition ships with same amount of fetcher helpers as enterprise
Free edition shipped with a reduced count of fetcher helpers, yet when updating
to the enterprice version this count remained locked and the enterprice system's
performance did not scale on higher load.

Thus the free edition ships the same amount of fetchers, which idle under the low
load of the free edition, but are able to work upon upgrade.
The ram usage of fetchers is not high anyway.

CMK-8122

Change-Id: I5789d6ef300d3d0b73c33860af67a8ba566760ca