13640 Deprecation of Web API
The WATO Web API is deprecated with this change. The Web API will continue to
work with all Checkmk 2.1 releases and removed with Checkmk 2.2.

We recommend to migrate all your existing scripts using the Web API to the REST
API as soon as possible to uncover issues ahead of time before moving to
Checkmk 2.2.

For users upgrading existing sites from 2.0 to 2.1 nothing will change.
However, sites created with 2.1 will be initialized with a disabled Web API by
default.

It is possible to enable or disable the Web API manually by using the global
setting "Site management > Disable Web API" depending on your needs.

Change-Id: I3e7344452ef6b7001e633de4e981251912323982