12704 FIX Do not create Carbon data if no connection is configured
The core would create and enqueue data for the Carbon connection
even if none is configured.  This may incur a small performance
penalty and, provided the site generates a large amount of data,
log messages such as
<tt>[carbon connection pool] queue full, didn't push update</tt>
that are irrelevant.

This fix only sends data to the Carbon connection if necessary.

SUP-9717

Change-Id: I7010aa0e62694e81516eb2091fa557ec5dd251b1