13619 FIX Fix crash due to dashboard's ill configured with software inventory object restriction
If you configured a dashboard with the object restriction "show information of a
single software package", that setting produced a misconfiguration of the
dashboard specification. When updating to checkmk 2.1 your site is unusable and
you can only see a crash report with the message:

TypeError ('NoneType' object is not iterable)

This werk fixes the crash, yet we still recommend to remove dashboards that use
that restriction.

Fix method return type for dynamically created class

VisualInfo.single_spec(self) -> List[Tuple[str, ValueSpec]]

Thus the dynamically generated class, which we can't typecheck must
respect this return value.

This bug is similar to what is fixed with 3364. Yet instead of having an
if clause to deal with None in cmk.gui.visuals.info_params better
respect the VisualInfo.single_spec return type

CMK-10007

Change-Id: Ibe122571f4d00f4bff017eb2b1397e967780dc4d