12705 Let SNMP accept config without IP address
Users will no be able to monitor SNMP devices without an address but
this lets them report an error correctly.

---

12705 FIX Fix "Checker and fetcher mismatch" for SNMP devices that cannot be resolved

The service Check_MK Discovery could produce a crash report with the
error "Checker and fetcher mismatch" if the address of an SNMP host
cannot be resolved.  Alternatively, the service would be OK although
the host was not checked.

Now, the service for the unreachable host will fail consistently (CRIT)
with the error: "Host has no IP address configured".

---

CMK-9761

Change-Id: Ia495b3bcbd34d582eaddf52414d16cef2ead87a9