13852 FIX Win-agent installer creates allow-pull-mode file correctly
Previously, custom action in MSI failed if the available Powershell was
too old. Such failure made impossible the obtaining of the previously
installed version of win-agent. For the use case <tt>update from 2.0 on
2.1</tt>, this means lack of connection between agent and site, i.e. stop
of the monitoring.

Since this release, the uninstaller custom action supports also
Powershell 4.0 thus solving the problem.

Change-Id: Ife69f68117ffcaec253fe94f9a9397a7b174686f