13790 FIX KUBE: Remove Container Labels
Previously, on a Kubernetes host Checkmk would create a single label of each

- cmk/container_image and

- cmk/container_name.

The label value was set to the name, respectively image, of one container. Since label keys
have to be unique, this would cause most containers to be left out, if there were multiple
containers.  This werk removes the faulty labels.

Change-Id: I606718738f6201b591b2d9d1139e69d1281a2b1c