Removed a performance pessimization regarding authorization.
There were a few broken attempts in the code to "improve" authorization
performance, which actually made things *much* worse in the extreme case: To
avoid a check which only costs O(1) per host, an authorization check was
left out under some circumstances.  The equivalent check was implictly done
later when the hosts were processed, but only *after* any filters have been
evaluated. If the filters do some non-trivial work, this is a bad tradeoff.

In general: Try to filter out things as soon as possible, this will give the
best complexity, and avoid premature "optimizations", which might turn out
to make things worse.

Change-Id: I872d5cfd14e3811add1997c720bade7d4aa3ea05