Extract make_sources
We have make_non_cluster_sources and make_cluster_sources, make_sources
is the generic version of these.  It uses either the cluster version or
the non-cluster version based on the configuration.

This lets us reduce the number of arguments to make_broker and somewhat
clarifies its API.

More importantly, the goal is to make sure that the sources correspond
to the fetcher messages in any case and this should best be realized
ASAP in the call stack.  So, a function may take the sources and fetch
the data from them but taking the messages and guessing their origin is
always risky.

CMK-10076

Change-Id: Iace76fb82d9d1babd6890320147d98f8c46bdbe7