Source success is independent of content of payload
This is not only a code simplification, but also
semantically correct: The success of the data source
does not depend on the content of the payload (or the mode).

It is not obvious to me what the "proper" solution for
piggyback would be -- but it is not examining the payload.

Change-Id: If0b3b5137b325750354e1aab62271bec756405ec