Clarify passing of FecherMessage for checking
One can see from the signatures of `commandline_checking()` and
`active_check_checking()` that the only difference between the two
functions is that `active_check_checking` takes the
`Sequence[FetcherMessage]` but `commandline_checking` does not.
Therefore, we make the argument non-optional for `active_check_checking`
and call the other function if the data is not ready yet.

This clarifies the call stack with regard to sources and
`fetcher_messages`.

CMK-10076

Change-Id: I52740876a16e2dfa2e8b58882f935a008cc6673f