Add 500ms delay between connect and read
This may(!) fix flakiness of the test_pull_ts. Or may not fix.

MOTIVATION: We have no right to read data immediately after the
connection and expect that data are ready. Especially in a heavy
loaded environment as our build Node or local VM. We must wait
a bit. Method to wait may differ. For integration tests simple
sleep is preferrable.

Change-Id: Ia12e35ee67ad82147cd8fc44cb3e365c4a4baa49