fields: MultiNested now supports merging
MultiNested represents a Union of two or more
Schemas. The problem of the previous approach
was that it could only represent

  "Either one or the other"

but it couldn't represent an object which had
the sum of all the keys of all objects and could
validate these keys on the sub-schemas.

This is useful in situations where e.g. some
attributes are defined statically, while others
are defined by user-data. Then you have multiple
schemas validating keys from the same object.

To achieve this, we first copy the input, then
try to decode it, then remove all sucessfully
converted keys and try again with the next sub-schema.

If no sub-schema yields any data, the operation is an error.

If one of the sub-schemas yields something, it's a success.

More than one sub-schema may yield data for an object.

This approach needs to be activated by setting "merged=True"
on instantiation of MultiNested. The default behaviour is
backwards compatible.

Change-Id: Ifbbb89a5079815f1b5d93b0e7dab034dd11026b6