Refactoring: Introduce a User object to bundle authorization code.
Although the change is quite big, it is completely mechanical and basically
easy: The main change is the introduction of the User class, which just
wraps existing code in auth.cc.  An instance of this is used instead of a
"naked" contact in the Table classes to simplify and centralize the
authorization checks.

Note that TableEventConsole::isAuthorizedForEvent() and Filter::accepts()
still use a "naked" contact instead of a User instance.  This will be fixed
in upcoming commits.

Change-Id: Ic4e101f2881d70f3f8d37f86e7d707ca4af63724