13677 FIX mk_apt: Check crashes on single-line agent sections
This fixes a regression that got introduced with Werk #13578.

While the above mentioned Werk fixed an issue with ignored security updates
for the "APT normal and security updates" plugin, it made a wrong assumption
to the output of the underlying <tt>apt-get</tt> command, that lead to a crash
for outputs that exactly consist of one security update item.

Change-Id: I8c60b1df1b0d72a5493112aaaffedd4924707a27