13419 FIX cisco_cpu_multiitem: fix Werk #13413
Werk #13413 in version 2.0.0p22 caused that many valid
`cisco_cpu_multiitem` services vanished. This werk brings them back, but
in some cases with a different name: If your `cisco_cpu_multiitem` is
named `CPU utilization 0` those might vanish and be rediscovered as
`CPU utilization 1`.

Werk #13413 assumed that a 0 in cpmCPUTotalPhysicalIndex means that this CPU is
not supported at all. This assumption was wrong.

Co-authored-by: Benedikt Seidl <benedikt.seidl@tribe29.com>
Change-Id: I8e99accb6501eee0d00c4fcd3f922aad40293b70