13926 FIX validation of error responses of the REST API
This werk changes the default error schema to be in line with
the until now returned responses.

The schema and responses did not fit and it was not found because
the responses were not checked automatically. This is now the case.

To ensure that dynamic scripts will not break only the schema has
been changed. Users of dynamic languages (Python, Bash, etc.) will not
have to change anything.

If you use a statically generated API client, you may need to re-compile
the client after this werk.

Change-Id: Icf9ac8d8bf694ff3b834712c9e1a7943149a3b33