Remove cycles between labels and ruleset_matcher
We need to chose one direction for the imports.  Since the manager
was not actually doing much with its attributes and methods, we
make it a simple NamedTuple and move the rest to the ruleset matcher.

Note that the ruleset match already handles tags, so it is not
surprising to have labels here as well.

CMK-10181

Change-Id: Ie6c38194598e357375ce8b55c685a6572d57cc42