13085 FIX REST API: fix missing host/folder attributes
When editing hosts or folders through the REST API there could be a case when
certain fields were not in the list of editable attributes. If a user attempted
to set this field, it would be treated as a "custom attribute" with a potentially
different data type. This would then lead to errors.

This werk fixes the disappearing attributes (e.g. locked_by, locked_attributes).

This werk also prevents internal attribute names to be used in "custom attributes".

There is no futher action to be taken by the user. It's in effect immediately.

Change-Id: Ifef3073b880fe7c7ba5331b9ce29604b488e1b25