Use None as the default for visibility.
The goal is to remove the none_value kwarg from cmk.valuespec.Optional: With
the current state of affairs, typing would be much more complicated than it
needs to be.  This was the last use site, finally!

Looking at the code, it seems to be the case that visibility values
("public") which are not a tuple are used in boolean contexts only, so it
doesn't matter if the value is False or None.

Change-Id: Ide1be4e13089c31dc2fa418580bc65298150483f