13887 Removed none_value keyword argument from Optional ValueSpec
Note: If you don't use the Optional ValueSpec in your checks or in your GUI
modifications, you are not affected by this change.  And even if you use it,
this change only affects you when you use the "none_value" keyword argument
with a value different from "None".  In a nutshell: Most users should not be
affected at all.

The "none_value" keyword argeument has been removed to improve the typing of
the Optional ValueSpec.  The previous usages of this were a bit error-prone
and not really necessary at all.  If you are affected by this removal, you
can use the Transform ValueSpec and/or tiny modifications of you code to
handle "None".

Change-Id: Ie8ecce71bba7cce8f4ac0bd7c2b54a40fec04dc9