13864 FIX Host registration: Enable on sites with disabled Setup & check site name
The host registration via the <tt>cmk-agent-ctl</tt> command failed for
sites with disabled Setup.

Furthermore, upon registration, we now check if the site where the host is
registering is also the site on which it is monitored (only relevant in
distributed setups). If not, the registration fails.

Finally, upon registration we also ensure that the host is not a cluster
host, since cluster hosts have no direct physical or virtual counterpart
which needs to be registered.

CMK-10258

Change-Id: I4ec9105ffa70badd17353cd1fe22b1dddaaef8dd