Notification spooler: Support for TLS authentication
Previously mknotifyd did not authenticate peers. With this Werk mknotifyd can
be configured to require TLS client certificate authentication. This is only
available if encryption for mknotifyd is configured. In order to verify the
peers correctly the corresponding site CAs must be trusted. This is usually the
case if you enabled the configuration sync and enabled the encryption of
livestatus. If you have another setup make sure both all Site CAs are trusted
on each site.

Change-Id: Icfb6c1446a4849b83b47582012942bd57a1033d0