Do not fetch live TCP data if not in checking mode
Fetching live data in a non-checking mode (e.g. discovery or inventory)
leads to data loss: information that is only transmitted once (logwatch
messages) is missed by the checks.

In order to avoid this, various measures are taken all over the code
base, to set caching options such that no live data is fetched.

With this patch we now radically disallow it. We expect this to lead to
some unpleasant crashes, but at least we'll see it then.

Change-Id: I2f5dd6372c63669e174b7d0766ebbc3834bd514d