13856 FIX Pipe symbol in name of the NIC doesn't lead to crash
In Windows names of a network card may contain a wide variety of the
symbols including the pipe symbol. If the pipe symbol is presented in
the name, then WATO crashes.

With this release, the windows_if.ps1 plugin replaces pipe symbol, i.e.
'|' with slash symbol, i.e. '/', thus eliminating the problem.

Ticket CMK-10291

Change-Id: Iedcdb221c487d425fad7f8ffdcf3f96fcfbe87ff