KUBE: inventory now shows Metadata instead of General
We debated this: Do we want to use the Metadata, which is more familiar
to Kubernetes users, but technically inappropriate, or General, which is
less descriptive, but more appropriate on a Cluster host?

Conclusion was that both are appropriate, so now we pick the one which
is more consistent with the inventory paths.

Change-Id: I662cdd748bb319bd7f191a8b5e2092fa6e764e8b