13928 FIX validation of custom tag-groups through the REST API
To create and edit custom tag groups via the REST API, one
had to restart Apache after a new tag was created and before
it could be used.

This werk fixes this behaviour. A restart is no longer necessary
for the tag-groups to be editable right after creation.

Note: The documentation of the custom created tags in the ReDoc component
are still only updated after an Apache restart. This does not affect the
editability of the tags though.

Change-Id: Ia0cd8fe314d0655a64c9364471239d098574fddb